requirements.txt
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
